create driver module

+++ smart_cab/src/main/java/com/example/smart_cab/service/DriverService.java
@@ -13,6 +13,10 @@ public class DriverService {
     @Autowired
     private DriverRepository driverRepository;
 
+    public Driver addDriver(Driver driver){
+        return driverRepository.save(driver);
+    }
+ 
     public List<Driver>getAllDrivers(){
         return driverRepository.findAll();
     }
@@ -25,21 +29,24 @@ public class DriverService {
         return driverRepository.save(driver);
     }
     
-    public Driver updateDriver(Long id,Driver driverDetails){
-        Driver driver=driverRepository.findById(id).orElseThrow(()-> new RuntimeException("Driver not found"));
-
-         if (driverDetails.getName() != null) driver.setName(driverDetails.getName());
-        if (driverDetails.getEmail() != null) driver.setEmail(driverDetails.getEmail());
-        if (driverDetails.getPhone() != null) driver.setPhone(driverDetails.getPhone());
-        if (driverDetails.getLicenseNumber() != null) driver.setLicenseNumber(driverDetails.getLicenseNumber());
-        if (driverDetails.getVehicleNumber() != null) driver.setVehicleNumber(driverDetails.getVehicleNumber());
-        driver.setAvailable(driverDetails.isAvailable());
-
-        return driverRepository.save(driver);
+    public Driver updateDriver(Long id, Driver updatedDriver) {
+        return driverRepository.findById(id).map(driver -> {
+            driver.setName(updatedDriver.getName());
+            driver.setPhone(updatedDriver.getPhone());
+            driver.setLicense_number(updatedDriver.getLicense_number());
+            driver.setRating(updatedDriver.getRating());
+            driver.setStatus(updatedDriver.getStatus());
+            return driverRepository.save(driver);
+        }).orElse(null);
     }
 
-    public void deleteDriver(Long id){
-        driverRepository.deleteById(id);
+
+    public boolean deleteDriver(Long id) {
+        if (driverRepository.existsById(id)) {
+            driverRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
